Fix compiler errors caused by uint128 on Windows platform

+++ src/log/gelf_appender.cpp
@@ -118,7 +118,7 @@ namespace fc
     gelf_message["_timestamp_ns"] = time_ns;
 
     static unsigned long gelf_log_counter;
-    gelf_message["_log_id"] = fc::to_string(++gelf_log_counter);
+    gelf_message["_log_id"] = fc::to_string((int)++gelf_log_counter);
 
     switch (context.get_log_level())
     {
